function moved from PipelineBuilder  to AppConfigService
